Move actions to actions file
