+++ examples/with-supabase/app/(auth-pages)/forgot-password/page.tsx
@@ -1,54 +1,15 @@
-import Link from "next/link";
-import { createClient } from "@/utils/supabase/server";
-import { redirect } from "next/navigation";
-import { SubmitButton } from "@/components/submit-button";
-import { Label } from "@/components/ui/label";
-import { Input } from "@/components/ui/input";
+import { forgotPassword } from "@/app/actions";
 import { FormMessage, Message } from "@/components/form-message";
-import { headers } from "next/headers";
-import { encodedRedirect } from "@/utils/utils";
+import { SubmitButton } from "@/components/submit-button";
+import { Input } from "@/components/ui/input";
+import { Label } from "@/components/ui/label";
+import Link from "next/link";
 
 export default function ForgotPassword({
   searchParams,
 }: {
   searchParams: Message;
 }) {
-  const forgotPassword = async (formData: FormData) => {
-    "use server";
-
-    const email = formData.get("email")?.toString();
-    const supabase = createClient();
-    const origin = headers().get("origin");
-    const callbackUrl = formData.get("callbackUrl")?.toString();
-
-    if (!email) {
-      return encodedRedirect("error", "/forgot-password", "Email is required");
-    }
-
-    const { error } = await supabase.auth.resetPasswordForEmail(email, {
-      redirectTo: `${origin}/auth/callback?redirect_to=/protected/reset-password`,
-    });
-
-    if (error) {
-      console.error(error.message);
-      return encodedRedirect(
-        "error",
-        "/forgot-password",
-        "Could not reset password",
-      );
-    }
-
-    if (callbackUrl) {
-      return redirect(callbackUrl);
-    }
-
-    return encodedRedirect(
-      "success",
-      "/forgot-password",
-      "Check your email for a link to reset your password.",
-    );
-  };
-
   return (
     <form className="flex-1 flex flex-col w-full gap-2 text-foreground [&>input]:mb-6 min-w-64 max-w-64">
       <div>
