+++ examples/with-supabase/app/(auth-pages)/protected/page.tsx
@@ -1,9 +1,9 @@
-import DeployButton from "@/components/deploy-dropdown";
 import AuthButton from "@/components/auth-button";
 import { createClient } from "@/utils/supabase/server";
 import FetchDataSteps from "@/components/tutorial/fetch-data-steps";
 import Header from "@/components/header";
 import { redirect } from "next/navigation";
+import DeployButton from "@/components/deploy-button";
 
 export default async function ProtectedPage() {
   const supabase = createClient();
