+++ examples/with-supabase/app/(auth-pages)/protected/reset-password/page.tsx
@@ -2,52 +2,13 @@ import { FormMessage, Message } from "@/components/form-message";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
 import { SubmitButton } from "@/components/submit-button";
-import { createClient } from "@/utils/supabase/server";
-import { encodedRedirect } from "@/utils/utils";
+import { resetPassword } from "@/app/actions";
 
 export default async function ResetPassword({
   searchParams,
 }: {
   searchParams: Message;
 }) {
-  const resetPassword = async (formData: FormData) => {
-    "use server";
-    const supabase = createClient();
-
-    const password = formData.get("password") as string;
-    const confirmPassword = formData.get("confirmPassword") as string;
-
-    if (!password || !confirmPassword) {
-      encodedRedirect(
-        "error",
-        "/protected/reset-password",
-        "Password and confirm password are required",
-      );
-    }
-
-    if (password !== confirmPassword) {
-      encodedRedirect(
-        "error",
-        "/protected/reset-password",
-        "Passwords do not match",
-      );
-    }
-
-    const { error } = await supabase.auth.updateUser({
-      password: password,
-    });
-
-    if (error) {
-      encodedRedirect(
-        "error",
-        "/protected/reset-password",
-        "Password update failed",
-      );
-    }
-
-    encodedRedirect("success", "/protected/reset-password", "Password updated");
-  };
-
   return (
     <div className="flex-1 flex flex-col items-center justify-center w-full">
       <form className="flex flex-col w-full max-w-md p-4 gap-2 [&>input]:mb-4">
